fix(alerts): keep every label set on the OTLP and JSON metrics paths
The remote-write path was fixed in #14010 to key its realtime-alert dedup
on the series signature, so a batch carrying many label sets contributes a
row each instead of only the first. OTLP and JSON still marked an alert
done on first match, so the same batch produced a single row there.

Move the dedup helpers from prom.rs into the metrics module so all three
paths share one implementation, and apply it to otlp.rs and json.rs. JSON
reuses the signature it already computes per record; OTLP recomputes it,
which is exact because HASH_LABEL is itself in the exclusion set.

Also folds in the review notes from #14010: single-line comments, and the
constant and type above the function block.

Refs #8696
